Created workspace activity and buttons.
Added a button on the main activity for transitioning into the workspace
activity. Created "Simulate", "Results", and "Save" buttons (save button
is disabled until logic is added to enable it). Want to get a navigation
drawer added to the workspace activity for displaying components. This is
the best idea I have for that right now. Jon's mockups are very generous
in the horizontal dimension...