Navigation drawer in place.
Added a functional navigation drawer that houses an expandable list view.
First level is the component categories (antennas, 2-port components,
1-port components) and the second level is the components under each
category. The categories collapse and expand as well (part of the
expandable list view). Next step is to attach listeners to the entries
that can later be used to drop down components or pull up the component
parameter modification window.