Created bare Component class.
Made bare minimum setter and getter functions for all the private
attributes. So far these are all the attributes I can think will be in any
given component. The idea is to have a single class instantiation be used
for any given component. The attributes should be suffient to describe the
component. May want to have some IDs in there later on to help identify it
(the name may be good enough, I suppose).