ComponentEditor starts and accepts object!
I only have it set up to launch the ComponentEditor activity when you long
click on Patch, but it successfully creates a patch component object and
sends it to the activity. The activity then populates a few fields from
the data in the component! Want to add class attributes to hold hints for
the parameters.