Units are now in a spinner.
ComponentEditor previously had the unit as a preset text, such as "mm" or
"GHz". This is now populated with ALL of the prefixes in the string-array
prefixes_abbrev concatenated with the appropriate unit for that component.
It now works gud! Also removed the now-unused strings and string-arrays
from strings.xml for stuff like kHz, mm, pF, and so on. Updated the
ComponentEditor layout a bit as well so it looks better with the spinners.