Added component thumbnail to ComponentEditor.
Had the ImageView already set up but did not have any images to put in
here. Now the proper image is pulled from a drawable resource and put in
here. Need to make an ImageView for the component's symbol and another for
the component's equation. A few thumbnails are missing in the DropBox
folder so I created an issue for this (and another issue for missing
component symbols).