Images in workspace with onClickListener.
Clicking a component in the drawer adds it to the workspace. Clicking a
component symbol in the workspace launches the component editor.
Components are (for now) placed directly below the component that was
most-recently-added to the workspace. Component symbol and thumbnail info
is now stored in each Component object, as these images are known at
compile time and can be stored upon Component creation.