Added three shunt components.
I want to think of a more graceful way for assigning component types
instead of a bunch of static ints. Enum would be GREAT but you cant just
get the int by using the dot operator, you gotta call a function
(ordinal() is already defined). Anyways, the new components are in there.