Made dialog for getting freq, sub parms.
Needed to prompt user to enter three pieces of info necessary for
simulating: center frequency, substrate height, substrate permittivity. I
made a DialogFragment which does this then populates the Circuit object
with this info. It's a bit kludgy and I want to make the callbacks more
elegant but it works for now. This way the guys can test the math and shit
with different frequencies, etc.